requirements.txt
boto3==1.9.111
google-auth-httplib2==0.0.3
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-oauthlib==0.3.0
python-dateutil==2.7.3
oauth2client==4.1.3
pytz==2018.4
PyJWT==1.7.1
